Rename package and module
The name was changed from nn-io-interact to blabot.
