+++ blabot/templated_io.py
